wire up mock patcher and navigate
